Conflicts:
	api/src/main/java/org/adbcj/support/DefaultDbFuture.java

+++ h2/src/main/java/org/adbcj/h2/packets/ParametersSerialisation.java
@@ -32,8 +32,8 @@ final class ParametersSerialisation {
             stream.writeInt(H2Types.LONG.id());
             stream.writeLong(((Long) param).longValue());
         } else if (Double.class.isInstance(param)) {
-            stream.writeInt(H2Types.LONG.id());
-            stream.writeDouble(((Double) param).longValue());
+            stream.writeInt(H2Types.DOUBLE.id());
+            stream.writeDouble(((Double) param).doubleValue());
         } else if (Date.class.isInstance(param)) {
             stream.writeInt(H2Types.DATE.id());
             stream.writeLong(DateTimeUtils.dateValueFromDate(((Date) param).getTime()));
